Force cURL to use TLSv1 for secure connections
On some distros cURL will use SSLv3 by default (e.g. on Debian7 with
the latest updates currently available). Many servers are blocking
secure connections that use SSLv3 to prevent heartbleed attack, and
you will get curl error #35. To prevent it, we have to force cURL to
use TLSv1, which it should use anyway. This is the newest version that
can safely be enforced, as TLSv1_0 and higher require curl 7.34+.